requirements.txt
Flask==3.1.2
langchain==1.2.7
langchain-openai==1.1.7
pydantic==2.12.5
python-dotenv==1.2.1
requests==2.32.5
yfinance==1.1.0
